// src/app/@core/data/models/registro-notas/docente-asignatura.ts
export class DocenteAsignatura {
    Docente: string;
    Identificacion: string;
    Codigo: string;
    Asignatura: string;
    Nivel: string;
    Grupo: string;
    Inscritos: number;
    Creditos: number;
    Periodo: string;
  }

/*   export interface DocenteAsignatura {
    Docente: string;
    Identificacion: string;
    Codigo: string;
    Asignatura: string;
    Nivel: string;
    Grupo: string;
    Inscritos: number;
    Creditos: number;
    Periodo: string;
  } */